@@ -15,15 +15,15 @@ import React from 'react';
 
 export default class App extends React.PureComponent { // eslint-disable-line react/prefer-stateless-function
 
-  static propTypes = {
-    children: React.PropTypes.node,
-  };
+    static propTypes = {
+        children: React.PropTypes.node,
+    };
 
-  render() {
-    return (
-      <div>
-        {React.Children.toArray(this.props.children)}
-      </div>
-    );
-  }
+    render() {
+        return (
+          <div>
+            {React.Children.toArray(this.props.children)}
+          </div>
+        );
+    }
 }
