@@ -55,30 +55,30 @@ const store = configureStore(initialState, browserHistory);
 // is under the non-default key ("routing"), selectLocationState
 // must be provided for resolving how to retrieve the "route" in the state
 const history = syncHistoryWithStore(browserHistory, store, {
-  selectLocationState: makeSelectLocationState(),
+    selectLocationState: makeSelectLocationState(),
 });
 
 // Set up the router, wrapping all Routes in the App component
 const rootRoute = {
-  component: App,
-  childRoutes: createRoutes(store),
+    component: App,
+    childRoutes: createRoutes(store),
 };
 
 const render = (messages) => {
-  ReactDOM.render(
-    <Provider store={store}>
-      <LanguageProvider messages={messages}>
-        <Router
-          history={history}
-          routes={rootRoute}
-          render={
+    ReactDOM.render(
+      <Provider store={store}>
+        <LanguageProvider messages={messages}>
+          <Router
+            history={history}
+            routes={rootRoute}
+            render={
             // Scroll to top when going to a new page, imitating default browser
             // behaviour
             applyRouterMiddleware(useScroll())
           }
-        />
-      </LanguageProvider>
-    </Provider>,
+          />
+        </LanguageProvider>
+      </Provider>,
     document.getElementById('app')
   );
 };
@@ -87,30 +87,30 @@ const render = (messages) => {
 if (module.hot) {
   // modules.hot.accept does not accept dynamic dependencies,
   // have to be constants at compile-time
-  module.hot.accept('./i18n', () => {
-    render(translationMessages);
-  });
+    module.hot.accept('./i18n', () => {
+        render(translationMessages);
+    });
 }
 
 // Chunked polyfill for browsers without Intl support
 if (!window.Intl) {
-  (new Promise((resolve) => {
-    resolve(import('intl'));
-  }))
+    (new Promise((resolve) => {
+        resolve(import('intl'));
+    }))
     .then(() => Promise.all([
-      import('intl/locale-data/jsonp/en.js'),
+        import('intl/locale-data/jsonp/en.js'),
     ]))
     .then(() => render(translationMessages))
     .catch((err) => {
-      throw err;
+        throw err;
     });
 } else {
-  render(translationMessages);
+    render(translationMessages);
 }
 
 // Install ServiceWorker and AppCache in the end since
 // it's not most important operation and if main code fails,
 // we do not want it installed
 if (process.env.NODE_ENV === 'production') {
-  require('offline-plugin/runtime').install(); // eslint-disable-line global-require
+    require('offline-plugin/runtime').install(); // eslint-disable-line global-require
 }
