@@ -1,18 +1,18 @@
 // makeSelectLocationState expects a plain JS object for the routing state
 const makeSelectLocationState = () => {
-  let prevRoutingState;
-  let prevRoutingStateJS;
+    let prevRoutingState;
+    let prevRoutingStateJS;
 
-  return (state) => {
-    const routingState = state.get('route'); // or state.route
+    return (state) => {
+        const routingState = state.get('route'); // or state.route
 
-    if (!routingState.equals(prevRoutingState)) {
-      prevRoutingState = routingState;
-      prevRoutingStateJS = routingState.toJS();
-    }
+        if (!routingState.equals(prevRoutingState)) {
+            prevRoutingState = routingState;
+            prevRoutingStateJS = routingState.toJS();
+        }
 
-    return prevRoutingStateJS;
-  };
+        return prevRoutingStateJS;
+    };
 };
 
 export {
