@@ -6,12 +6,12 @@ import HomePage from '../index';
 import messages from '../messages';
 
 describe('<HomePage />', () => {
-  it('should render the page message', () => {
-    const renderedComponent = shallow(
-      <HomePage />
+    it('should render the page message', () => {
+        const renderedComponent = shallow(
+          <HomePage />
     );
-    expect(renderedComponent.contains(
-      <FormattedMessage {...messages.header} />
+        expect(renderedComponent.contains(
+          <FormattedMessage {...messages.header} />
     )).toEqual(true);
-  });
+    });
 });
