@@ -9,8 +9,8 @@ import {
 } from './constants';
 
 export function changeLocale(languageLocale) {
-  return {
-    type: CHANGE_LOCALE,
-    locale: languageLocale,
-  };
+    return {
+        type: CHANGE_LOCALE,
+        locale: languageLocale,
+    };
 }
