@@ -14,17 +14,17 @@ import {
 } from '../App/constants'; // eslint-disable-line
 
 const initialState = fromJS({
-  locale: DEFAULT_LOCALE,
+    locale: DEFAULT_LOCALE,
 });
 
 function languageProviderReducer(state = initialState, action) {
-  switch (action.type) {
-    case CHANGE_LOCALE:
-      return state
+    switch (action.type) {
+                    case CHANGE_LOCALE:
+                        return state
         .set('locale', action.locale);
-    default:
-      return state;
-  }
+                    default:
+                        return state;
+    }
 }
 
 export default languageProviderReducer;
