@@ -7,13 +7,13 @@ import {
 } from '../constants';
 
 describe('LanguageProvider actions', () => {
-  describe('Change Local Action', () => {
-    it('has a type of CHANGE_LOCALE', () => {
-      const expected = {
-        type: CHANGE_LOCALE,
-        locale: 'de',
-      };
-      expect(changeLocale('de')).toEqual(expected);
+    describe('Change Local Action', () => {
+        it('has a type of CHANGE_LOCALE', () => {
+            const expected = {
+                type: CHANGE_LOCALE,
+                locale: 'de',
+            };
+            expect(changeLocale('de')).toEqual(expected);
+        });
     });
-  });
 });
