@@ -10,40 +10,40 @@ import configureStore from '../../../store';
 import { translationMessages } from '../../../i18n';
 
 const messages = defineMessages({
-  someMessage: {
-    id: 'some.id',
-    defaultMessage: 'This is some default message',
-    en: 'This is some en message',
-  },
+    someMessage: {
+        id: 'some.id',
+        defaultMessage: 'This is some default message',
+        en: 'This is some en message',
+    },
 });
 
 describe('<LanguageProvider />', () => {
-  it('should render its children', () => {
-    const children = (<h1>Test</h1>);
-    const renderedComponent = shallow(
-      <LanguageProvider messages={messages} locale="en">
-        {children}
-      </LanguageProvider>
+    it('should render its children', () => {
+        const children = (<h1>Test</h1>);
+        const renderedComponent = shallow(
+          <LanguageProvider messages={messages} locale="en">
+            {children}
+          </LanguageProvider>
     );
-    expect(renderedComponent.contains(children)).toBe(true);
-  });
+        expect(renderedComponent.contains(children)).toBe(true);
+    });
 });
 
 describe('<ConnectedLanguageProvider />', () => {
-  let store;
+    let store;
 
-  beforeAll(() => {
-    store = configureStore({}, browserHistory);
-  });
+    beforeAll(() => {
+        store = configureStore({}, browserHistory);
+    });
 
-  it('should render the default language messages', () => {
-    const renderedComponent = mount(
-      <Provider store={store}>
-        <ConnectedLanguageProvider messages={translationMessages}>
-          <FormattedMessage {...messages.someMessage} />
-        </ConnectedLanguageProvider>
-      </Provider>
+    it('should render the default language messages', () => {
+        const renderedComponent = mount(
+          <Provider store={store}>
+            <ConnectedLanguageProvider messages={translationMessages}>
+              <FormattedMessage {...messages.someMessage} />
+            </ConnectedLanguageProvider>
+          </Provider>
     );
-    expect(renderedComponent.contains(<FormattedMessage {...messages.someMessage} />)).toBe(true);
-  });
+        expect(renderedComponent.contains(<FormattedMessage {...messages.someMessage} />)).toBe(true);
+    });
 });
