@@ -5,11 +5,11 @@ import {
 } from '../selectors';
 
 describe('selectLanguage', () => {
-  it('should select the global state', () => {
-    const globalState = fromJS({});
-    const mockedState = fromJS({
-      language: globalState,
+    it('should select the global state', () => {
+        const globalState = fromJS({});
+        const mockedState = fromJS({
+            language: globalState,
+        });
+        expect(selectLanguage(mockedState)).toEqual(globalState);
     });
-    expect(selectLanguage(mockedState)).toEqual(globalState);
-  });
 });
