@@ -15,11 +15,11 @@ import { FormattedMessage } from 'react-intl';
 import messages from './messages';
 
 export default class NotFound extends React.PureComponent { // eslint-disable-line react/prefer-stateless-function
-  render() {
-    return (
-      <h1>
-        <FormattedMessage {...messages.header} />
-      </h1>
-    );
-  }
+    render() {
+        return (
+          <h1>
+            <FormattedMessage {...messages.header} />
+          </h1>
+        );
+    }
 }
