@@ -11,24 +11,24 @@ import { DEFAULT_LOCALE } from './containers/App/constants'; // eslint-disable-l
 import enTranslationMessages from './translations/en.json';
 
 export const appLocales = [
-  'en',
+    'en',
 ];
 
 addLocaleData(enLocaleData);
 
 export const formatTranslationMessages = (locale, messages) => {
-  const defaultFormattedMessages = locale !== DEFAULT_LOCALE
+    const defaultFormattedMessages = locale !== DEFAULT_LOCALE
     ? formatTranslationMessages(DEFAULT_LOCALE, enTranslationMessages)
     : {};
-  return Object.keys(messages).reduce((formattedMessages, key) => {
-    let message = messages[key];
-    if (!message && locale !== DEFAULT_LOCALE) {
-      message = defaultFormattedMessages[key];
-    }
-    return Object.assign(formattedMessages, { [key]: message });
-  }, {});
+    return Object.keys(messages).reduce((formattedMessages, key) => {
+        let message = messages[key];
+        if (!message && locale !== DEFAULT_LOCALE) {
+            message = defaultFormattedMessages[key];
+        }
+        return Object.assign(formattedMessages, { [key]: message });
+    }, {});
 };
 
 export const translationMessages = {
-  en: formatTranslationMessages('en', enTranslationMessages),
+    en: formatTranslationMessages('en', enTranslationMessages),
 };
