@@ -19,31 +19,31 @@ import languageProviderReducer from 'containers/LanguageProvider/reducer';
 
 // Initial routing state
 const routeInitialState = fromJS({
-  locationBeforeTransitions: null,
+    locationBeforeTransitions: null,
 });
 
 /**
  * Merge route into the global application state
  */
 function routeReducer(state = routeInitialState, action) {
-  switch (action.type) {
+    switch (action.type) {
     /* istanbul ignore next */
-    case LOCATION_CHANGE:
-      return state.merge({
-        locationBeforeTransitions: action.payload,
-      });
-    default:
-      return state;
-  }
+                    case LOCATION_CHANGE:
+                        return state.merge({
+                            locationBeforeTransitions: action.payload,
+                        });
+                    default:
+                        return state;
+    }
 }
 
 /**
  * Creates the main reducer with the asynchronously loaded ones
  */
 export default function createReducer(asyncReducers) {
-  return combineReducers({
-    route: routeReducer,
-    language: languageProviderReducer,
-    ...asyncReducers,
-  });
+    return combineReducers({
+        route: routeReducer,
+        language: languageProviderReducer,
+        ...asyncReducers,
+    });
 }
