@@ -6,21 +6,21 @@ import { browserHistory } from 'react-router';
 import configureStore from '../store';
 
 describe('configureStore', () => {
-  let store;
+    let store;
 
-  beforeAll(() => {
-    store = configureStore({}, browserHistory);
-  });
-
-  describe('asyncReducers', () => {
-    it('should contain an object for async reducers', () => {
-      expect(typeof store.asyncReducers).toBe('object');
+    beforeAll(() => {
+        store = configureStore({}, browserHistory);
     });
-  });
 
-  describe('runSaga', () => {
-    it('should contain a hook for `sagaMiddleware.run`', () => {
-      expect(typeof store.runSaga).toBe('function');
+    describe('asyncReducers', () => {
+        it('should contain an object for async reducers', () => {
+            expect(typeof store.asyncReducers).toBe('object');
+        });
     });
-  });
+
+    describe('runSaga', () => {
+        it('should contain a hook for `sagaMiddleware.run`', () => {
+            expect(typeof store.runSaga).toBe('function');
+        });
+    });
 });
