@@ -9,75 +9,75 @@
 const componentExists = require('../utils/componentExists');
 
 module.exports = {
-  description: 'Add an unconnected component',
-  prompts: [{
-    type: 'list',
-    name: 'type',
-    message: 'Select the type of component',
-    default: 'Stateless Function',
-    choices: () => ['Stateless Function', 'ES6 Class (Pure)', 'ES6 Class'],
-  }, {
-    type: 'input',
-    name: 'name',
-    message: 'What should it be called?',
-    default: 'Button',
-    validate: (value) => {
-      if ((/.+/).test(value)) {
-        return componentExists(value) ? 'A component or container with this name already exists' : true;
-      }
-
-      return 'The name is required';
-    },
-  }, {
-    type: 'confirm',
-    name: 'wantMessages',
-    default: true,
-    message: 'Do you want i18n messages (i.e. will this component use text)?',
-  }],
-  actions: (data) => {
-    // Generate index.js and index.test.js
-    let componentTemplate;
-
-    switch (data.type) {
-      case 'ES6 Class': {
-        componentTemplate = './component/es6.js.hbs';
-        break;
-      }
-      case 'ES6 Class (Pure)': {
-        componentTemplate = './component/es6.pure.js.hbs';
-        break;
-      }
-      case 'Stateless Function': {
-        componentTemplate = './component/stateless.js.hbs';
-        break;
-      }
-      default: {
-        componentTemplate = './component/es6.js.hbs';
-      }
-    }
-
-    const actions = [{
-      type: 'add',
-      path: '../../app/components/{{properCase name}}/index.js',
-      templateFile: componentTemplate,
-      abortOnFail: true,
+    description: 'Add an unconnected component',
+    prompts: [{
+        type: 'list',
+        name: 'type',
+        message: 'Select the type of component',
+        default: 'Stateless Function',
+        choices: () => ['Stateless Function', 'ES6 Class (Pure)', 'ES6 Class'],
     }, {
-      type: 'add',
-      path: '../../app/components/{{properCase name}}/tests/index.test.js',
-      templateFile: './component/test.js.hbs',
-      abortOnFail: true,
-    }];
+        type: 'input',
+        name: 'name',
+        message: 'What should it be called?',
+        default: 'Button',
+        validate: (value) => {
+            if ((/.+/).test(value)) {
+                return componentExists(value) ? 'A component or container with this name already exists' : true;
+            }
+
+            return 'The name is required';
+        },
+    }, {
+        type: 'confirm',
+        name: 'wantMessages',
+        default: true,
+        message: 'Do you want i18n messages (i.e. will this component use text)?',
+    }],
+    actions: (data) => {
+    // Generate index.js and index.test.js
+        let componentTemplate;
+
+        switch (data.type) {
+                        case 'ES6 Class': {
+                            componentTemplate = './component/es6.js.hbs';
+                            break;
+                        }
+                        case 'ES6 Class (Pure)': {
+                            componentTemplate = './component/es6.pure.js.hbs';
+                            break;
+                        }
+                        case 'Stateless Function': {
+                            componentTemplate = './component/stateless.js.hbs';
+                            break;
+                        }
+                        default: {
+                            componentTemplate = './component/es6.js.hbs';
+                        }
+        }
+
+        const actions = [{
+            type: 'add',
+            path: '../../app/components/{{properCase name}}/index.js',
+            templateFile: componentTemplate,
+            abortOnFail: true,
+        }, {
+            type: 'add',
+            path: '../../app/components/{{properCase name}}/tests/index.test.js',
+            templateFile: './component/test.js.hbs',
+            abortOnFail: true,
+        }];
 
     // If they want a i18n messages file
-    if (data.wantMessages) {
-      actions.push({
-        type: 'add',
-        path: '../../app/components/{{properCase name}}/messages.js',
-        templateFile: './component/messages.js.hbs',
-        abortOnFail: true,
-      });
-    }
+        if (data.wantMessages) {
+            actions.push({
+                type: 'add',
+                path: '../../app/components/{{properCase name}}/messages.js',
+                templateFile: './component/messages.js.hbs',
+                abortOnFail: true,
+            });
+        }
 
-    return actions;
-  },
+        return actions;
+    },
 };
