@@ -12,17 +12,17 @@ const routeGenerator = require('./route/index.js');
 const languageGenerator = require('./language/index.js');
 
 module.exports = (plop) => {
-  plop.setGenerator('component', componentGenerator);
-  plop.setGenerator('container', containerGenerator);
-  plop.setGenerator('route', routeGenerator);
-  plop.setGenerator('language', languageGenerator);
-  plop.addHelper('directory', (comp) => {
-    try {
-      fs.accessSync(path.join(__dirname, `../../app/containers/${comp}`), fs.F_OK);
-      return `containers/${comp}`;
-    } catch (e) {
-      return `components/${comp}`;
-    }
-  });
-  plop.addHelper('curly', (object, open) => (open ? '{' : '}'));
+    plop.setGenerator('component', componentGenerator);
+    plop.setGenerator('container', containerGenerator);
+    plop.setGenerator('route', routeGenerator);
+    plop.setGenerator('language', languageGenerator);
+    plop.addHelper('directory', (comp) => {
+        try {
+            fs.accessSync(path.join(__dirname, `../../app/containers/${comp}`), fs.F_OK);
+            return `containers/${comp}`;
+        } catch (e) {
+            return `components/${comp}`;
+        }
+    });
+    plop.addHelper('curly', (object, open) => (open ? '{' : '}'));
 };
