@@ -11,7 +11,7 @@ const pageContainers = fs.readdirSync(path.join(__dirname, '../../../app/contain
 const components = pageComponents.concat(pageContainers);
 
 function componentExists(comp) {
-  return components.indexOf(comp) >= 0;
+    return components.indexOf(comp) >= 0;
 }
 
 module.exports = componentExists;
