@@ -1,3 +1,4 @@
 // needed for regenerator-runtime
 // (ES7 generator support is required by redux-saga)
-import 'babel-polyfill';
+// import 'babel-polyfill';
+import 'jest-enzyme';
