@@ -13,50 +13,50 @@ const pkg = require(path.resolve(process.cwd(), 'package.json'));
 const dllPlugin = pkg.dllPlugin;
 
 const plugins = [
-  new webpack.HotModuleReplacementPlugin(), // Tell webpack we want hot reloading
-  new webpack.NoErrorsPlugin(),
-  new HtmlWebpackPlugin({
-    inject: true, // Inject all files that are generated by webpack, e.g. bundle.js
-    templateContent: templateContent(), // eslint-disable-line no-use-before-define
-  }),
-  new CircularDependencyPlugin({
-    exclude: /a\.js|node_modules/, // exclude node_modules
-    failOnError: false, // show a warning when there is a circular dependency
-  }),
+    new webpack.HotModuleReplacementPlugin(), // Tell webpack we want hot reloading
+    new webpack.NoErrorsPlugin(),
+    new HtmlWebpackPlugin({
+        inject: true, // Inject all files that are generated by webpack, e.g. bundle.js
+        templateContent: templateContent(), // eslint-disable-line no-use-before-define
+    }),
+    new CircularDependencyPlugin({
+        exclude: /a\.js|node_modules/, // exclude node_modules
+        failOnError: false, // show a warning when there is a circular dependency
+    }),
 ];
 
 module.exports = require('./webpack.base.babel')({
   // Add hot reloading in development
-  entry: [
-    'eventsource-polyfill', // Necessary for hot reloading with IE
-    'webpack-hot-middleware/client?reload=true',
-    path.join(process.cwd(), 'app/app.js'), // Start with js/app.js
-  ],
+    entry: [
+        'eventsource-polyfill', // Necessary for hot reloading with IE
+        'webpack-hot-middleware/client?reload=true',
+        path.join(process.cwd(), 'app/app.js'), // Start with js/app.js
+    ],
 
   // Don't use hashes in dev mode for better performance
-  output: {
-    filename: '[name].js',
-    chunkFilename: '[name].chunk.js',
-  },
+    output: {
+        filename: '[name].js',
+        chunkFilename: '[name].chunk.js',
+    },
 
   // Add development plugins
-  plugins: dependencyHandlers().concat(plugins), // eslint-disable-line no-use-before-define
+    plugins: dependencyHandlers().concat(plugins), // eslint-disable-line no-use-before-define
 
   // Tell babel that we want to hot-reload
-  babelQuery: {
+    babelQuery: {
     // require.resolve solves the issue of relative presets when dealing with
     // locally linked packages. This is an issue with babel and webpack.
     // See https://github.com/babel/babel-loader/issues/149 and
     // https://github.com/webpack/webpack/issues/1866
-    presets: ['babel-preset-react-hmre'].map(require.resolve),
-  },
+        presets: ['babel-preset-react-hmre'].map(require.resolve),
+    },
 
   // Emit a source map for easier debugging
-  devtool: 'cheap-module-eval-source-map',
+    devtool: 'cheap-module-eval-source-map',
 
-  performance: {
-    hints: false,
-  },
+    performance: {
+        hints: false,
+    },
 });
 
 /**
@@ -70,61 +70,61 @@ module.exports = require('./webpack.base.babel')({
  */
 function dependencyHandlers() {
   // Don't do anything during the DLL Build step
-  if (process.env.BUILDING_DLL) { return []; }
+    if (process.env.BUILDING_DLL) { return []; }
 
   // If the package.json does not have a dllPlugin property, use the CommonsChunkPlugin
-  if (!dllPlugin) {
-    return [
-      new webpack.optimize.CommonsChunkPlugin({
-        name: 'vendor',
-        children: true,
-        minChunks: 2,
-        async: true,
-      }),
-    ];
-  }
+    if (!dllPlugin) {
+        return [
+            new webpack.optimize.CommonsChunkPlugin({
+                name: 'vendor',
+                children: true,
+                minChunks: 2,
+                async: true,
+            }),
+        ];
+    }
 
-  const dllPath = path.resolve(process.cwd(), dllPlugin.path || 'node_modules/react-boilerplate-dlls');
+    const dllPath = path.resolve(process.cwd(), dllPlugin.path || 'node_modules/react-boilerplate-dlls');
 
   /**
    * If DLLs aren't explicitly defined, we assume all production dependencies listed in package.json
    * Reminder: You need to exclude any server side dependencies by listing them in dllConfig.exclude
    */
-  if (!dllPlugin.dlls) {
-    const manifestPath = path.resolve(dllPath, 'reactBoilerplateDeps.json');
+    if (!dllPlugin.dlls) {
+        const manifestPath = path.resolve(dllPath, 'reactBoilerplateDeps.json');
 
-    if (!fs.existsSync(manifestPath)) {
-      logger.error('The DLL manifest is missing. Please run `npm run build:dll`');
-      process.exit(0);
+        if (!fs.existsSync(manifestPath)) {
+            logger.error('The DLL manifest is missing. Please run `npm run build:dll`');
+            process.exit(0);
+        }
+
+        return [
+            new webpack.DllReferencePlugin({
+                context: process.cwd(),
+                manifest: require(manifestPath), // eslint-disable-line global-require
+            }),
+        ];
     }
-
-    return [
-      new webpack.DllReferencePlugin({
-        context: process.cwd(),
-        manifest: require(manifestPath), // eslint-disable-line global-require
-      }),
-    ];
-  }
 
   // If DLLs are explicitly defined, we automatically create a DLLReferencePlugin for each of them.
-  const dllManifests = Object.keys(dllPlugin.dlls).map((name) => path.join(dllPath, `/${name}.json`));
+    const dllManifests = Object.keys(dllPlugin.dlls).map((name) => path.join(dllPath, `/${name}.json`));
 
-  return dllManifests.map((manifestPath) => {
-    if (!fs.existsSync(path)) {
-      if (!fs.existsSync(manifestPath)) {
-        logger.error(`The following Webpack DLL manifest is missing: ${path.basename(manifestPath)}`);
-        logger.error(`Expected to find it in ${dllPath}`);
-        logger.error('Please run: npm run build:dll');
+    return dllManifests.map((manifestPath) => {
+        if (!fs.existsSync(path)) {
+            if (!fs.existsSync(manifestPath)) {
+                logger.error(`The following Webpack DLL manifest is missing: ${path.basename(manifestPath)}`);
+                logger.error(`Expected to find it in ${dllPath}`);
+                logger.error('Please run: npm run build:dll');
 
-        process.exit(0);
-      }
-    }
+                process.exit(0);
+            }
+        }
 
-    return new webpack.DllReferencePlugin({
-      context: process.cwd(),
-      manifest: require(manifestPath), // eslint-disable-line global-require
+        return new webpack.DllReferencePlugin({
+            context: process.cwd(),
+            manifest: require(manifestPath), // eslint-disable-line global-require
+        });
     });
-  });
 }
 
 /**
@@ -132,17 +132,17 @@ function dependencyHandlers() {
  * DLL Javascript files are loaded in script tags and available to our application.
  */
 function templateContent() {
-  const html = fs.readFileSync(
+    const html = fs.readFileSync(
     path.resolve(process.cwd(), 'app/index.html')
   ).toString();
 
-  if (!dllPlugin) { return html; }
+    if (!dllPlugin) { return html; }
 
-  const doc = cheerio(html);
-  const body = doc.find('body');
-  const dllNames = !dllPlugin.dlls ? ['reactBoilerplateDeps'] : Object.keys(dllPlugin.dlls);
+    const doc = cheerio(html);
+    const body = doc.find('body');
+    const dllNames = !dllPlugin.dlls ? ['reactBoilerplateDeps'] : Object.keys(dllPlugin.dlls);
 
-  dllNames.forEach((dllName) => body.append(`<script data-dll='true' src='/${dllName}.dll.js'></script>`));
+    dllNames.forEach((dllName) => body.append(`<script data-dll='true' src='/${dllName}.dll.js'></script>`));
 
-  return doc.toString();
+    return doc.toString();
 }
